// src/app/models/vendor/vendor.models.ts
export interface IVendor {
    vendorName: string;
    vendorDescription: string;
    vendorType: string;
    addressLine1: string;
    addressLine2: string;
    city: string;
    adminEmail: string;
    telephone1: string;
    telephone2: string;
    facebookLink: string;
    twitterLink: string;
    instagramLink: string;
}